wip issue naemspace link with nostr note :(
