+++ scripts/package.json
@@ -33,6 +33,7 @@
     "relay:dev": "nodemon --ext *"
   },
   "dependencies": {
+    "@noble/hashes": "^1.2.1",
     "@noble/secp256k1": "^1.7.1",
     "@nostr-dev-kit/ndk": "^2.10.1",
     "afk_prototype": "link:",
@@ -44,8 +45,7 @@
     "nostr-tools": "^2.5.2",
     "path": "^0.12.7",
     "starknet": "^6.11.0",
-    "ws": "^8.17.0",
-    "@noble/hashes": "^1.2.1"
+    "ws": "^8.17.0"
   },
   "devDependencies": {
     "@types/chai": "^4.3.16",
